Moved files + merge policy system base
