+++ src/main/java/net/mcreator/generator/DefinitionsProvider.java
@@ -22,7 +22,7 @@ import net.mcreator.element.BaseType;
 import net.mcreator.element.ModElementType;
 import net.mcreator.element.ModElementTypeLoader;
 import net.mcreator.plugin.PluginLoader;
-import net.mcreator.util.YamlMerge;
+import net.mcreator.util.yaml.YamlMerge;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.snakeyaml.engine.v2.exceptions.YamlEngineException;
