+++ src/main/java/net/mcreator/minecraft/DataListLoader.java
@@ -20,7 +20,7 @@ package net.mcreator.minecraft;
 
 import net.mcreator.io.FileIO;
 import net.mcreator.plugin.PluginLoader;
-import net.mcreator.util.YamlUtil;
+import net.mcreator.util.yaml.YamlUtil;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.snakeyaml.engine.v2.api.Load;
