+++ src/main/java/net/mcreator/util/yaml/YamlMerge.java
@@ -17,13 +17,14 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.mcreator.util;
+package net.mcreator.util.yaml;
 
 import net.mcreator.io.FileIO;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.snakeyaml.engine.v2.api.Load;
 
+import javax.annotation.Nullable;
 import java.net.URL;
 import java.util.*;
 
@@ -32,6 +33,10 @@ public class YamlMerge {
 	private static final Logger LOG = LogManager.getLogger(YamlMerge.class);
 
 	public static Map<?, ?> multiLoadYAML(Enumeration<URL> resources) {
+		return multiLoadYAML(resources, null);
+	}
+
+	public static Map<?, ?> multiLoadYAML(Enumeration<URL> resources, @Nullable MergePolicy policy) {
 		List<Map<?, ?>> loadedMaps = new ArrayList<>();
 
 		Collections.list(resources).forEach(resource -> {
@@ -42,16 +47,20 @@ public class YamlMerge {
 			}
 		});
 
-		return deepMerge(loadedMaps);
+		if (policy == null)
+			policy = MergePolicy.DEFAULT;
+
+		return deepMerge(loadedMaps, policy);
 	}
 
-	private static Map<Object, Object> deepMerge(List<Map<?, ?>> maps) {
+	private static Map<Object, Object> deepMerge(List<Map<?, ?>> maps, MergePolicy policy) {
 		Map<Object, Object> result = new HashMap<>();
-		maps.forEach(map -> mergeInto(result, map));
+		maps.forEach(map -> mergeInto(result, map, policy));
 		return result;
 	}
 
-	@SuppressWarnings("unchecked") private static void mergeInto(Map<Object, Object> target, Map<?, ?> source) {
+	@SuppressWarnings("unchecked")
+	private static void mergeInto(Map<Object, Object> target, Map<?, ?> source, MergePolicy policy) {
 		for (Map.Entry<?, ?> entry : source.entrySet()) {
 			Object key = entry.getKey();
 			Object value = entry.getValue();
@@ -65,21 +74,43 @@ public class YamlMerge {
 
 			if (existing instanceof Map && value instanceof Map) {
 				try {
-					mergeInto((Map<Object, Object>) existing, (Map<?, ?>) value);
+					mergeInto((Map<Object, Object>) existing, (Map<?, ?>) value, policy);
 				} catch (Throwable t) {
 					LOG.warn("Failed to merge key {}", key, t);
 				}
 			} else if (existing instanceof List && value instanceof List) {
 				try {
-					// merge lists so that new entries are added on the top, since most
-					// processors use the last match in MCreator's yaml files as the final result
-					((List<Object>) existing).addAll(0, (List<Object>) value);
+					policy.mergeList(key, (List<Object>) existing, (List<Object>) value);
+				} catch (Throwable t) {
+					LOG.warn("Failed to merge key {}", key, t);
+				}
+			} else {
+				try {
+					policy.mergeScalar(key, existing, value);
 				} catch (Throwable t) {
 					LOG.warn("Failed to merge key {}", key, t);
 				}
 			}
-			// No need to check scalars, scalar already exists -> first wins
 		}
+	}
+
+	public interface MergePolicy {
+
+		MergePolicy DEFAULT = new MergePolicy() {};
+
+		default void mergeList(Object key, List<Object> existing, List<Object> addition) throws Exception {
+			try {
+				// merge lists so that new entries are added on the top, since most
+				// processors use the last match in MCreator's yaml files as the final result
+				existing.addAll(0, addition);
+			} catch (Throwable t) {
+				throw new Exception(t);
+			}
+		}
+
+		default void mergeScalar(Object key, Object existing, Object addition) throws Exception {
+		}
+
 	}
 
 }
